+++ tests/testthat/test-lazy-select-query.R
@@ -1,7 +1,7 @@
 test_that("can print lazy_select_query", {
   expect_snapshot(
     lazy_select_query(
-      from = lazy_query_local(tibble(x = 1, y = 2), "df"),
+      x = lazy_query_local(tibble(x = 1, y = 2), "df"),
       last_op = "select",
       select = quos(
         x_mean = mean(x),